Move PortQueue attribute helpers to a neutral utils home
config copp queue reused utils::applyPortQueueConfig by including a header
under commands/config/qos/, coupling the copp command tree to the qos one.
The parser and applier operate on cfg::PortQueue and know nothing about
qos, so move QueueIdAndAttributes, validQueueAttrs, and
applyPortQueueConfig to fboss/cli/fboss2/utils/PortQueueConfigUtils.
The genuinely qos-specific helpers (QueueConfigName and the named
queue-config list lookups) stay behind, renamed to QueueConfigUtils so
the two headers cannot be confused.
