+++ fboss/cli/fboss2/BUCK
@@ -1174,8 +1174,8 @@ cpp_library(
         "commands/config/ptp/CmdConfigPtp.cpp",
         "commands/config/ptp/transparent_clock/CmdConfigPtpTransparentClock.cpp",
         "commands/config/qos/CmdConfigQos.cpp",
-        "commands/config/qos/PortQueueConfigUtils.cpp",
         "commands/config/qos/QosPolicyUtils.cpp",
+        "commands/config/qos/QueueConfigUtils.cpp",
         "commands/config/qos/buffer_pool/CmdConfigQosBufferPool.cpp",
         "commands/config/qos/default_policy/CmdConfigQosDefaultPolicy.cpp",
         "commands/config/qos/policy/CmdConfigQosPolicy.cpp",
@@ -1248,6 +1248,7 @@ cpp_library(
         "session/Git.cpp",
         "session/SystemdInterface.cpp",
         "utils/InterfaceList.cpp",
+        "utils/PortQueueConfigUtils.cpp",
     ],
     headers = [
         "commands/config/CmdConfigAppliedInfo.h",
@@ -1330,8 +1331,8 @@ cpp_library(
         "commands/config/ptp/CmdConfigPtp.h",
         "commands/config/ptp/transparent_clock/CmdConfigPtpTransparentClock.h",
         "commands/config/qos/CmdConfigQos.h",
-        "commands/config/qos/PortQueueConfigUtils.h",
         "commands/config/qos/QosPolicyUtils.h",
+        "commands/config/qos/QueueConfigUtils.h",
         "commands/config/qos/buffer_pool/CmdConfigQosBufferPool.h",
         "commands/config/qos/default_policy/CmdConfigQosDefaultPolicy.h",
         "commands/config/qos/policy/CmdConfigQosPolicy.h",
@@ -1404,6 +1405,7 @@ cpp_library(
         "session/Git.h",
         "session/SystemdInterface.h",
         "utils/InterfaceList.h",
+        "utils/PortQueueConfigUtils.h",
     ],
     deps = [
         "fbsource//third-party/glog:glog",
