+++ fboss/cli/fboss2/test/config/CmdDeleteQosQueueConfigTest.cpp
@@ -13,7 +13,7 @@
 #include <string>
 #include <vector>
 
-#include "fboss/cli/fboss2/commands/config/qos/PortQueueConfigUtils.h"
+#include "fboss/cli/fboss2/commands/config/qos/QueueConfigUtils.h"
 #include "fboss/cli/fboss2/commands/delete/qos/queue_config/CmdDeleteQosQueueConfig.h"
 #include "fboss/cli/fboss2/commands/delete/qos/queue_config/CmdDeleteQosQueueConfigQueueId.h"
 #include "fboss/cli/fboss2/session/ConfigSession.h"
