+++ fboss/cli/fboss2/utils/PortQueueConfigUtils.h
@@ -20,32 +20,6 @@
 
 namespace facebook::fboss::utils {
 
-// Reserved queue-config name designating the switch-wide default queue list.
-// `queue-config default` edits SwitchConfig::defaultPortQueues -- the fallback
-// every port without an explicit portQueueConfigName uses -- while any other
-// name edits a SwitchConfig::portQueueConfigs entry. Reserving the name is what
-// keeps the two disjoint: portQueueConfigs can never acquire a "default" key,
-// so Port::portQueueConfigName can never resolve to the default list.
-inline constexpr auto kDefaultQueueConfigName = "default";
-
-/**
- * Name of a queue config: either a portQueueConfigs key or the reserved
- * `default`.
- */
-class QueueConfigName : public BaseObjectArgType<std::string> {
- public:
-  // NOLINTNEXTLINE(google-explicit-constructor)
-  /* implicit */ QueueConfigName(std::vector<std::string> v);
-
-  const std::string& getName() const {
-    return data_[0];
-  }
-
-  bool isDefault() const {
-    return data_[0] == kDefaultQueueConfigName;
-  }
-};
-
 /**
  * A queue id plus the attributes to apply to it.
  *
@@ -83,33 +57,10 @@ class QueueIdAndAttributes : public BaseObjectArgType<std::string> {
 // text and the parser's error messages so the two cannot drift.
 const std::string& validQueueAttrs();
 
-// Resolves `name` to the PortQueue list it designates, creating an empty
-// portQueueConfigs entry if a named config does not exist yet.
-std::vector<cfg::PortQueue>& queueConfigListForWrite(
-    cfg::SwitchConfig& switchConfig,
-    const QueueConfigName& name);
-
-// Resolves `name` to the PortQueue list it designates, or nullptr if a named
-// config does not exist. Callers that must not create an entry on a typo --
-// delete, in particular -- use this rather than queueConfigListForWrite.
-// `default` always resolves, since defaultPortQueues always exists.
-std::vector<cfg::PortQueue>* findQueueConfigList(
-    cfg::SwitchConfig& switchConfig,
-    const QueueConfigName& name);
-
-// Names of the ports whose Port::portQueueConfigName references `name`, so a
-// caller can refuse to remove a queue config that is still bound and would
-// leave those ports pointing at nothing.
-//
-// Always empty for the reserved `default`: no port names it explicitly, and
-// every port without a binding falls back to it, so it can never be dangling.
-std::vector<std::string> portsUsingQueueConfig(
-    const cfg::SwitchConfig& switchConfig,
-    const QueueConfigName& name);
-
 // Applies queue configuration attributes onto `queue`. Shared by every command
-// in the `config qos queue-config <name|default>` family, which configure the
-// same cfg::PortQueue attributes into different parts of the switch config.
+// that configures cfg::PortQueue attributes -- the `config qos queue-config
+// <name|default>` family and `config copp queue` -- which write the same
+// struct into different parts of the switch config.
 //
 // `attributes` is the flat list of scalar <attr, value> pairs (reserved-bytes,
 // shared-bytes, weight, scaling-factor, scheduling, stream-type,
